refactor(sample): split DemoSections into per-section files, data-driven tabs
- Split 923-line DemoSections.kt into 7 focused files under sections/ package:
  StylingSection, TypographySection, TogglesSection, CallbacksSection,
  ThemeCreationSection, AdvancedEngineSection, EngineInfoSection
- Shared SectionHeader/SubSectionHeader moved to sections/SectionComponents.kt
- Shared KOTLIN_SNIPPET/PYTHON_SNIPPET moved to SampleData.kt as internal vals
- Add DemoTab sealed class — adding a tab now requires one change, not two
- Name magic color fallbacks in AdvancedEngineSection (DarkCodeBackground etc.)
- Replace fixed-height nested LazyColumn in EngineInfoSection with plain Column
