+++ sample/src/main/kotlin/dev/hossain/highlight/sample/sections/SectionComponents.kt
@@ -1,0 +1,53 @@
+package dev.hossain.highlight.sample.sections
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+/**
+ * A simple section label rendered above each code block.
+ *
+ * Displays the language [title] in uppercase to visually separate samples in the list.
+ */
+@Composable
+internal fun SectionHeader(title: String) {
+    Row(
+        modifier =
+            Modifier
+                .fillMaxWidth()
+                .padding(bottom = 4.dp),
+        horizontalArrangement = Arrangement.SpaceBetween,
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        Text(text = title.uppercase())
+    }
+}
+
+/**
+ * A smaller sub-section label for grouping related items within a demo section.
+ */
+@Composable
+internal fun SubSectionHeader(title: String) {
+    Text(
+        text = title,
+        style =
+            TextStyle(
+                fontSize = 12.sp,
+                fontWeight = FontWeight.Medium,
+                color = MaterialTheme.colorScheme.onSurfaceVariant,
+                fontFamily = FontFamily.SansSerif,
+            ),
+        modifier = Modifier.padding(bottom = 4.dp),
+    )
+}
